pyproject.toml
[project]
name = "bertopic-pipeline"
version = "0.1.0"
description = "BERTopic topic modeling pipeline with CLI interface"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bertopic>=0.16.0",
    "sentence-transformers>=3.0.0",
    "umap-learn>=0.5.0",
    "hdbscan>=0.8.33",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "scrapy>=2.14.2",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]
viz = [
    "wordcloud>=1.9.0",
    "matplotlib>=3.7.0",
]
zhihu = [
    "playwright>=1.40.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
